+++ tests/benchmark/util.go
@@ -1,8 +1,11 @@
 package main
 
 import (
+	"bytes"
 	"io"
+	"log"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
 )
@@ -44,4 +47,22 @@ func copyDirExcept(src, dest, excludeSuffix string) error {
 		_, err = io.Copy(destFile, srcFile)
 		return err
 	})
+}
+
+
+func run(dir string, args ...string) ([]byte, error) {
+	stdoutBuf := bytes.Buffer{}
+	log.Printf("running %s in %s", strings.Join(args, " "), dir)
+	cmd := exec.Command(args[0], args[1:]...)
+	cmd.Stdout = &stdoutBuf
+	cmd.Stderr = os.Stderr
+	cmd.Dir = dir
+	err := cmd.Run()
+	if err != nil {
+		if exitErr, ok := err.(*exec.ExitError); ok {
+			log.Printf("stderr: %s", string(exitErr.Stderr))
+		}
+		return nil, err
+	}
+	return stdoutBuf.Bytes(), nil
 }
